CHET-120: Move failed file migration to it's own class
