+++ src/main/java/com/atlassian/migration/datacenter/spi/fs/reporting/FileSystemMigrationErrorReport.java
@@ -1,8 +1,5 @@
 package com.atlassian.migration.datacenter.spi.fs.reporting;
 
-import com.atlassian.migration.datacenter.core.fs.reporting.DefaultFileSystemMigrationErrorReport;
-
-import java.nio.file.Path;
 import java.util.List;
 
 /**
@@ -12,24 +9,6 @@ public interface FileSystemMigrationErrorReport {
 
     List<FailedFileMigration> getFailedFiles();
 
-    void reportFileNotMigrated(DefaultFileSystemMigrationErrorReport.FailedFileMigration failedFileMigration);
+    void reportFileNotMigrated(FailedFileMigration failedFileMigration);
 
-    class FailedFileMigration {
-        private Path filePath;
-
-        private String reason;
-
-        public FailedFileMigration(Path filePath, String reason) {
-            this.filePath = filePath;
-            this.reason = reason;
-        }
-
-        public Path getFilePath() {
-            return filePath;
-        }
-
-        public String getReason() {
-            return reason;
-        }
-    }
 }
